code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aom=3.6.0=h6a678d5_0
audioread=3.0.1=py312h7900ff3_2
blas=1.0=mkl
bottleneck=1.4.2=py312ha883a20_0
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
brotli-python=1.1.0=py312h2ec8cdc_2
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.1.31=hbcca054_0
cairo=1.16.0=hb05425b_5
certifi=2024.12.14=py312h06a4308_0
cffi=1.17.1=py312h06ac9bb_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
contourpy=1.3.1=py312h68727a3_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.8=3
cycler=0.12.1=pyhd8ed1ab_1
cyrus-sasl=2.1.28=h52b45da_1
dav1d=1.2.1=h5eee18b_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pyhd8ed1ab_1
expat=2.6.4=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
fontconfig=2.14.2=h14ed4e7_0
fonttools=4.55.8=py312h178313f_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.12.0=py312h06a4308_0
gettext=0.22.5=he02047a_3
gettext-tools=0.22.5=he02047a_3
giflib=5.2.2=h5eee18b_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py312h5eee18b_0
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h2=4.1.0=pyhd8ed1ab_1
harfbuzz=4.3.0=hf52aaf7_2
hpack=4.1.0=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=73.1=h6a678d5_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.6.1=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.5=py312h06a4308_0
joblib=1.4.2=pyhd8ed1ab_1
jpeg=9e=h5eee18b_3
kiwisolver=1.4.8=py312h84d6215_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy-loader=0.4=pyhd8ed1ab_2
lazy_loader=0.4=pyhd8ed1ab_2
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
leptonica=1.82.0=h42c8aad_2
lerc=4.0.0=h6a678d5_0
libabseil=20240116.2=cxx17_h6a678d5_0
libarchive=3.7.7=hfab0078_0
libasprintf=0.22.5=he8f35ee_3
libasprintf-devel=0.22.5=he8f35ee_3
libblas=3.9.0=1_h86c2bf4_netlib
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=8_h3b12eaf_netlib
libclang=14.0.6=default_hc6dbbc7_2
libclang13=14.0.6=default_he11475f_2
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.13.0.11=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.9.55=0
libcurl=8.11.1=hc9e6f67_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.22=h5eee18b_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libexpat=2.6.4=h5888daf_0
libffi=3.4.4=h6a678d5_1
libflac=1.4.3=h59595ed_0
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgettextpo=0.22.5=he02047a_3
libgettextpo-devel=0.22.5=he02047a_3
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libglib=2.78.4=hdc74915_0
libgomp=14.2.0=h77fa898_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=8_h3b12eaf_netlib
libllvm14=14.0.6=hcd5def8_4
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.8.0.86=0
libnsl=2.0.1=hd590300_0
libnvjpeg=11.9.0.86=0
libogg=1.3.5=h27cfd23_1
libopus=1.3.1=h5eee18b_1
libpng=1.6.39=h5eee18b_0
libpq=17.2=hdbd6064_0
libprotobuf=4.25.3=he621ea3_0
librosa=0.10.2.post1=pyhd8ed1ab_1
libsndfile=1.2.2=hc60ed4a_1
libsqlite=3.46.0=hde9e2c9_0
libssh2=1.11.1=h251f7ec_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libtasn1=4.19.0=h5eee18b_0
libtheora=1.1.1=h7f8727e_3
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libvorbis=1.3.7=h7b6447c_0
libvpx=1.13.1=h6a678d5_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.5=hfdd30dd_0
libzlib=1.2.13=h4ab18f5_6
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.43.0=py312h6a678d5_1
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py312h5eee18b_1
matplotlib=3.10.0=py312h06a4308_0
matplotlib-base=3.10.0=py312hd3ec401_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_2
mkl_fft=1.3.11=py312h5eee18b_0
mkl_random=1.2.8=py312h526ad5a_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpg123=1.32.9=hc50e24c_0
mpmath=1.3.0=py312h06a4308_0
msgpack-python=1.1.0=py312h68727a3_0
munkres=1.1.4=pyh9f0ad1d_0
mysql=8.4.0=h29a9f33_1
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.4.2=py312h06a4308_0
numba=0.60.0=py312h83e6fd3_0
numexpr=2.10.1=py312h3c60e43_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openldap=2.6.4=h42fbc30_0
openssl=3.4.0=h7b32b05_1
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py312h6a678d5_0
pcre2=10.42=hebb0a14_1
pillow=11.1.0=py312hcea889d_0
pip=25.0=py312h06a4308_0
pixman=0.40.0=h7f8727e_1
platformdirs=4.3.6=pyhd8ed1ab_1
ply=3.11=py312h06a4308_1
pooch=1.8.2=pyhd8ed1ab_1
pycparser=2.22=pyh29332c3_1
pyparsing=3.2.1=pyhd8ed1ab_0
pyqt=5.15.10=py312h6a678d5_1
pyqt5-sip=12.13.0=py312h5eee18b_1
pysocks=1.7.1=pyha55dd90_7
pysoundfile=0.13.1=pyhd8ed1ab_0
python=3.12.2=hab00c5b_0_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.12=5_cp312
pytorch=2.5.1=py3.12_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8=h7e8668a_6
pytorch-mutex=1.0=cuda
pytz=2024.1=py312h06a4308_0
pyyaml=6.0.2=py312h5eee18b_0
qhull=2020.2=h434a139_5
qt-main=5.15.2=hb6262e9_11
readline=8.2=h5eee18b_0
requests=2.32.3=pyhd8ed1ab_1
scikit-learn=1.6.1=py312h7a48858_0
scipy=1.15.1=py312h180e4f1_0
setuptools=75.8.0=py312h06a4308_0
sip=6.7.12=py312h6a678d5_1
six=1.17.0=pyhd8ed1ab_0
soxr=0.1.3=h0b41bf4_3
soxr-python=0.5.0.post1=py312h2ec8cdc_1
sqlite=3.45.3=h5eee18b_0
sympy=1.13.3=py312h06a4308_1
tbb=2021.8.0=hdb19cb5_0
tesseract=5.2.0=h6a678d5_2
threadpoolctl=3.5.0=pyhc1e730c_0
tk=8.6.14=h39e8969_0
torchtriton=3.1.0=py312
torchvision=0.20.1=py312_cu118
tornado=6.4.2=py312h5eee18b_0
tqdm=4.66.5=py312he106c6f_0
typing_extensions=4.12.2=py312h06a4308_0
tzdata=2025a=h04d1e81_0
unicodedata2=16.0.0=py312h66e93f0_0
urllib3=2.3.0=pyhd8ed1ab_0
wheel=0.45.1=py312h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h4ab18f5_6
zstandard=0.23.0=py312hef9b889_1
zstd=1.5.6=hc292b87_0
 